+++ js09/index.html
@@ -55,6 +55,7 @@
             </main>
         </div>
         <script src="./menu.js" type="module"></script>
+        <script src="./switchTheme.js" type="module"></script>
         <script src="./script.js" type="module"></script>
     </body>
 </html>
